WIP: Document and refactor filesystem

+++ src/os/kernel/filesystem/RamFs/PciNode.cc
@@ -15,7 +15,7 @@ uint64_t PciNode::getLength() {
     return cache.length();
 }
 
-char *PciNode::readData(char *buf, uint64_t pos, uint32_t numBytes) {
+char *PciNode::readData(char *buf, uint64_t pos, uint64_t numBytes) {
 
     if (cache.isEmpty()) {
         cacheDeviceList();
@@ -32,7 +32,7 @@ char *PciNode::readData(char *buf, uint64_t pos, uint32_t numBytes) {
     return buf;
 }
 
-int32_t PciNode::writeData(char *buf, uint64_t pos, uint32_t numBytes) {
+int64_t PciNode::writeData(char *buf, uint64_t pos, uint64_t numBytes) {
     return -1;
 }
