+++ src/os/kernel/filesystem/RamFs/RandomNode.cc
@@ -9,7 +9,7 @@ uint64_t RandomNode::getLength() {
     return 0;
 }
 
-char *RandomNode::readData(char *buf, uint64_t pos, uint32_t numBytes) {
+char *RandomNode::readData(char *buf, uint64_t pos, uint64_t numBytes) {
     auto *random = new Random(0xFF);
     
     for(uint32_t i = 0; i < numBytes; i++) {
@@ -20,6 +20,6 @@ char *RandomNode::readData(char *buf, uint64_t pos, uint32_t numBytes) {
     return buf;
 }
 
-int32_t RandomNode::writeData(char *buf, uint64_t pos, uint32_t numBytes) {
+int64_t RandomNode::writeData(char *buf, uint64_t pos, uint64_t numBytes) {
     return -1;
 }
