+++ src/os/kernel/filesystem/RamFs/ZeroNode.cc
@@ -8,11 +8,11 @@ uint64_t ZeroNode::getLength() {
     return 0;
 }
 
-char *ZeroNode::readData(char *buf, uint64_t pos, uint32_t numBytes) {
+char *ZeroNode::readData(char *buf, uint64_t pos, uint64_t numBytes) {
     memset(buf, 0, numBytes);
     return buf;
 }
 
-int32_t ZeroNode::writeData(char *buf, uint64_t pos, uint32_t numBytes) {
+int64_t ZeroNode::writeData(char *buf, uint64_t pos, uint64_t numBytes) {
     return -1;
 }
